requirements.txt
azure-storage-blob==12.20.0
openai==1.59.7
requests==2.32.3
beautifulsoup4==4.12.3
langchain==0.3.14
python-dotenv==1.0.1